[MP-5173] 신규 탭바 추가 개발중
