+++ DealiDesignSystemSampleApp/DealiDesignSystemSampleApp/ViewControllers/MainViewController.swift
@@ -150,20 +150,6 @@ final class MainViewController: UIViewController {
             $0.addTarget(self, action: #selector(tagButtonPressed), for: .touchUpInside)
         }
         
-        let tabBarComponents = DealiControl.btnOutlineLarge03()
-        contentStackView.addArrangedSubview(tabBarComponents)
-        tabBarComponents.do {
-            $0.title = "TabBar Components"
-            $0.addTarget(self, action: #selector(tabBarComponentsPressed), for: .touchUpInside)
-        }
-        
-        let tabBarV2Components = DealiControl.btnOutlineLarge03()
-        contentStackView.addArrangedSubview(tabBarV2Components)
-        tabBarV2Components.do {
-            $0.title = "TabBar v2 Components"
-            $0.addTarget(self, action: #selector(tabBarV2ComponentsPressed), for: .touchUpInside)
-        }
-        
         let checkComponents = DealiControl.btnOutlineLarge03()
         contentStackView.addArrangedSubview(checkComponents)
         checkComponents.do {
@@ -258,14 +244,6 @@ extension MainViewController {
     
     @objc func tagButtonPressed() {
         self.navigationController?.pushViewController(TagViewController(), animated: true)
-    }
-    
-    @objc func tabBarComponentsPressed() {
-        self.navigationController?.pushViewController(TabBarComponentViewController(), animated: true)
-    }
-    
-    @objc func tabBarV2ComponentsPressed() {
-        self.navigationController?.pushViewController(TabBarV2ComponentViewController(), animated: true)
     }
     
     @objc func checkComponentsPressed() {
