+++ DealiDesignSystemSampleApp/DealiDesignSystemSampleApp/ViewControllers/Sub/TabBarViewController.swift
@@ -26,18 +26,27 @@ class TabBarViewController: UIViewController {
                                  DealiTabBarItem.make(title: "8번 Tab"),
                                  DealiTabBarItem.make(title: "9번 Tab")]
     
+    private var segmentTabBarItems: [DealiTabBarItem] = []
+    private var tabBarSlider01Items: [DealiTabBarItem] = []
+    private var tabBarSlider02Items: [DealiTabBarItem] = []
+    
     private var tabBarSegment01ViewController: DealiTabBarViewController?
     private var tabBarSlider01ViewController: DealiTabBarViewController?
     private var tabBarSlider02ViewController: DealiTabBarViewController?
     
-    private let tabBarSegment01 = DealiTabBar_v2.tabBarSegment01()
-    private let tabBarSlider01 = DealiTabBar_v2.tabBarSlider01()
-    private let tabBarSlider02 = DealiTabBar_v2.tabBarSlider02()
-    private let tabBarChip01 = DealiTabBar_v2.tabBarChip01()
+    private let tabBarSegment01 = DealiTabBar.tabBarSegment01()
+    private let tabBarSlider01 = DealiTabBar.tabBarSlider01()
+    private let tabBarSlider02 = DealiTabBar.tabBarSlider02()
+    private let tabBarChip01 = DealiTabBar.tabBarChip01()
     
     private let tabBarSegment01ContentView = UIView()
     private let tabBarSlider01ContentView = UIView()
     private let tabBarSlider02ContentView = UIView()
+    
+    private let badgeOnOffButton = DealiControl.btnOutlineLarge01()
+    private let moveTabBatItemButton = DealiControl.btnOutlineLarge01()
+    private let hiddenTabBatItemButton = DealiControl.btnOutlineLarge01()
+    private let changeTitleTabBatItemButton = DealiControl.btnOutlineLarge01()
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -46,19 +55,38 @@ class TabBarViewController: UIViewController {
         self.view.backgroundColor = .white
         
         self.setTabBar()
-        
-//        self.tabBarView02.setTabBarItems(tabBarItemArray: sliderTabBarItemArray, startIndex: 1, isStandAloneView: true)
     }
     
     override func loadView() {
         super.loadView()
+        
+        let buttonContainerView = UIView()
+        self.view.addSubview(buttonContainerView)
+        buttonContainerView.then {
+            $0.backgroundColor = .white
+        }.snp.makeConstraints {
+            $0.left.right.bottom.equalToSuperview()
+        }
+        
+        let buttonContainerStackView = UIStackView()
+        buttonContainerView.addSubview(buttonContainerStackView)
+        buttonContainerStackView.then {
+            $0.axis = .vertical
+            $0.spacing =  10.0
+            $0.alignment = .center
+            $0.distribution = .equalSpacing
+        }.snp.makeConstraints {
+            $0.top.equalToSuperview()
+            $0.bottom.left.right.equalToSuperview()
+        }
         
         let contentScrollView = UIScrollView()
         self.view.addSubview(contentScrollView)
         contentScrollView.then {
             $0.bounces = false
         }.snp.makeConstraints {
-            $0.top.bottom.left.right.equalToSuperview()
+            $0.top.left.right.equalToSuperview()
+            $0.bottom.equalTo(buttonContainerView.snp.top)
         }
         
         let contentView = UIView()
@@ -180,10 +208,43 @@ class TabBarViewController: UIViewController {
             $0.left.right.equalToSuperview()
             $0.height.equalTo(300.0)
         }
+        
+        buttonContainerStackView.addArrangedSubview(self.badgeOnOffButton)
+        self.badgeOnOffButton.then {
+            $0.title = "뱃지 Off"
+            $0.isSelected = false
+            $0.addTarget(self, action: #selector(badgeOnOffButtonPressed), for: .touchUpInside)
+        }.snp.makeConstraints {
+            $0.left.right.equalToSuperview().inset(20.0)
+        }
+        
+        buttonContainerStackView.addArrangedSubview(self.moveTabBatItemButton)
+        self.moveTabBatItemButton.then {
+            $0.title = "특정 탭으로 이동"
+            $0.addTarget(self, action: #selector(moveTabBatItemButtonPressed), for: .touchUpInside)
+        }.snp.makeConstraints {
+            $0.left.right.equalToSuperview().inset(20.0)
+        }
+        
+        buttonContainerStackView.addArrangedSubview(self.hiddenTabBatItemButton)
+        self.hiddenTabBatItemButton.then {
+            $0.title = "특정 탭 숨김"
+            $0.isSelected = false
+            $0.addTarget(self, action: #selector(hiddenTabBatItemButtonPressed), for: .touchUpInside)
+        }.snp.makeConstraints {
+            $0.left.right.equalToSuperview().inset(20.0)
+        }
+        
+        buttonContainerStackView.addArrangedSubview(self.changeTitleTabBatItemButton)
+        self.changeTitleTabBatItemButton.then {
+            $0.title = "특정 탭 Title 변경"
+            $0.addTarget(self, action: #selector(changeTitleTabBatItemButtonPressed), for: .touchUpInside)
+        }.snp.makeConstraints {
+            $0.left.right.equalToSuperview().inset(20.0)
+        }
     }
 
     private func setTabBar() {
-        var segmentTabBarItems: [DealiTabBarItem] = []
         for i in 0..<3 {
             let viewController = DealiTabBarChildViewController()
             viewController.view.backgroundColor = self.randomColor()
@@ -191,58 +252,55 @@ class TabBarViewController: UIViewController {
             if i == 1 {
                 
                 let tabBarItem = DealiTabBarItem.make(viewController, title: "DETAILS", isHidden: false, showsBadge: false, icon: DealiTabBarIcon(url: URL(string: "https://v4.img.sinsang.market?f=https://image-cache.sinsang.market/home_tab/img_mbs_filled_16_ver01.png&w=48&h=48"), size: CGSize(width: 48.0, height: 48.0)))
-                segmentTabBarItems.append(tabBarItem)
+                self.segmentTabBarItems.append(tabBarItem)
             } else {
                 let tabBarItem = DealiTabBarItem.make(viewController, title: "\(i)번 Tab")
-                segmentTabBarItems.append(tabBarItem)
+                self.segmentTabBarItems.append(tabBarItem)
             }
         }
         
-        
-        var tabBarSlider01Items: [DealiTabBarItem] = []
         for i in 0..<10 {
             let viewController = DealiTabBarChildViewController()
             viewController.view.backgroundColor = self.randomColor()
             if i == 3 {
                 
                 let tabBarItem = DealiTabBarItem.make(viewController, title: "DETAILS", isHidden: false, showsBadge: false, icon: DealiTabBarIcon(url: URL(string: "https://v4.img.sinsang.market?f=https://image-cache.sinsang.market/home_tab/img_mbs_filled_16_ver01.png&w=48&h=48"), size: CGSize(width: 48.0, height: 48.0)))
-                tabBarSlider01Items.append(tabBarItem)
+                self.tabBarSlider01Items.append(tabBarItem)
             } else {
                 let tabBarItem = DealiTabBarItem.make(viewController, title: "\(i)번 Tab")
-                tabBarSlider01Items.append(tabBarItem)
+                self.tabBarSlider01Items.append(tabBarItem)
             }
         }
         
-        var tabBarSlider02Items: [DealiTabBarItem] = []
         for i in 0..<10 {
             let viewController = DealiTabBarChildViewController()
             viewController.view.backgroundColor = self.randomColor()
             if i == 3 {
                 
                 let tabBarItem = DealiTabBarItem.make(viewController, title: "DETAILS", isHidden: false, showsBadge: false, icon: DealiTabBarIcon(url: URL(string: "https://v4.img.sinsang.market?f=https://image-cache.sinsang.market/home_tab/img_mbs_filled_16_ver01.png&w=48&h=48"), size: CGSize(width: 48.0, height: 48.0)))
-                tabBarSlider02Items.append(tabBarItem)
+                self.tabBarSlider02Items.append(tabBarItem)
             } else {
                 let tabBarItem = DealiTabBarItem.make(viewController, title: "\(i)번 Tab")
-                tabBarSlider02Items.append(tabBarItem)
+                self.tabBarSlider02Items.append(tabBarItem)
             }
         }
 
-        let tabBarSegment01View = DealiTabBar_v2.tabBarSegment01()
-        self.tabBarSegment01ViewController = DealiTabBarViewController(tabBarView: tabBarSegment01View, tabBarItemArray: segmentTabBarItems)
+        let tabBarSegment01View = DealiTabBar.tabBarSegment01()
+        self.tabBarSegment01ViewController = DealiTabBarViewController(tabBarView: tabBarSegment01View, tabBarItemArray: self.segmentTabBarItems)
         self.tabBarSegment01ViewController?.startPageIndex = 7
         if let tabBarSegment01ViewController = self.tabBarSegment01ViewController {
             self.insertChildController(tabBarSegment01ViewController, intoParentView: self.tabBarSegment01ContentView)
         }
         
-        let tabBarSlider01View = DealiTabBar_v2.tabBarSlider01()
-        self.tabBarSlider01ViewController = DealiTabBarViewController(tabBarView: tabBarSlider01View, tabBarItemArray: tabBarSlider01Items)
+        let tabBarSlider01View = DealiTabBar.tabBarSlider01()
+        self.tabBarSlider01ViewController = DealiTabBarViewController(tabBarView: tabBarSlider01View, tabBarItemArray: self.tabBarSlider01Items)
         self.tabBarSlider01ViewController?.startPageIndex = 1
         if let tabBarSlider01ViewController = self.tabBarSlider01ViewController {
             self.insertChildController(tabBarSlider01ViewController, intoParentView: self.tabBarSlider01ContentView)
         }
         
-        let tabBarSlider02View = DealiTabBar_v2.tabBarSlider02()
-        self.tabBarSlider02ViewController = DealiTabBarViewController(tabBarView: tabBarSlider02View, tabBarItemArray: tabBarSlider02Items)
+        let tabBarSlider02View = DealiTabBar.tabBarSlider02()
+        self.tabBarSlider02ViewController = DealiTabBarViewController(tabBarView: tabBarSlider02View, tabBarItemArray: self.tabBarSlider02Items)
         self.tabBarSlider02ViewController?.startPageIndex = 5
         if let tabBarSlider02ViewController = self.tabBarSlider02ViewController {
             self.insertChildController(tabBarSlider02ViewController, intoParentView: self.tabBarSlider02ContentView)
@@ -258,8 +316,85 @@ class TabBarViewController: UIViewController {
 
 }
 
-extension TabBarViewController: DealiTabBarViewDelegate_v2 {
-    func didSelectTabBar(_ tabbarView: DealiTabBarView_v2, selectedIndex index: Int, showScrollAnimation animation: Bool) {
+extension TabBarViewController {
+    @objc func badgeOnOffButtonPressed() {
+        print("badgeOnOffButtonPressed")
+        self.badgeOnOffButton.isSelected.toggle()
+        
+        self.badgeOnOffButton.title = (self.badgeOnOffButton.isSelected == true ? "뱃지 On" : "뱃지 off")
+        
+        self.tabBarSegment01ViewController?.showTabBarItemBadge(index: 2, shouldShowBadge: (self.badgeOnOffButton.isSelected == true ? true : false))
+        self.tabBarSlider01ViewController?.showTabBarItemBadge(index: 2, shouldShowBadge: (self.badgeOnOffButton.isSelected == true ? true : false))
+        self.tabBarSlider02ViewController?.showTabBarItemBadge(index: 2, shouldShowBadge: (self.badgeOnOffButton.isSelected == true ? true : false))
+        
+        self.tabBarSegment01.showTabBarItemBadge(index: 2, shouldShowBadge: (self.badgeOnOffButton.isSelected == true ? true : false))
+        self.tabBarSlider01.showTabBarItemBadge(index: 2, shouldShowBadge: (self.badgeOnOffButton.isSelected == true ? true : false))
+        self.tabBarSlider02.showTabBarItemBadge(index: 2, shouldShowBadge: (self.badgeOnOffButton.isSelected == true ? true : false))
+        
+        
+    }
+    
+    @objc func moveTabBatItemButtonPressed() {
+        self.moveTabBatItemButton.isSelected.toggle()
+        
+        self.moveTabBatItemButton.title = (self.moveTabBatItemButton.isSelected == true ? "Tab으로 이동" : "0번 Tab으로 이동")
+        
+        self.tabBarSegment01ViewController?.moveViewerPageIndex(index: (self.moveTabBatItemButton.isSelected == true ? 1 : 0), animation: true)//setSelectedIndex(index: 1)
+        self.tabBarSlider01ViewController?.moveViewerPageIndex(index: (self.moveTabBatItemButton.isSelected == true ? 6 : 0), animation: false)
+        self.tabBarSlider02ViewController?.moveViewerPageIndex(index: (self.moveTabBatItemButton.isSelected == true ? 4 : 0), animation: true)
+        
+        self.tabBarSegment01.setSelectedIndex(index: (self.moveTabBatItemButton.isSelected == true ? 2 : 0))
+        self.tabBarSlider01.setSelectedIndex(index: (self.moveTabBatItemButton.isSelected == true ? 4 : 0))
+        self.tabBarSlider02.setSelectedIndex(index: (self.moveTabBatItemButton.isSelected == true ? 3 : 0))
+        self.tabBarChip01.setSelectedIndex(index: (self.moveTabBatItemButton.isSelected == true ? 5 : 0))
+    }
+    
+    @objc func hiddenTabBatItemButtonPressed() {
+        self.hiddenTabBatItemButton.isSelected.toggle()
+        
+        self.hiddenTabBatItemButton.title = (self.hiddenTabBatItemButton.isSelected == true ? "Tab 노출" : "Tab 숨김")
+        
+        self.segmentTabBarItemArray[1].isHidden = (self.hiddenTabBatItemButton.isSelected == true ? true : false)
+        self.sliderTabBarItemArray[4].isHidden = (self.hiddenTabBatItemButton.isSelected == true ? true : false)
+        
+        self.segmentTabBarItems[1].isHidden = (self.hiddenTabBatItemButton.isSelected == true ? true : false)
+        self.tabBarSlider01Items[4].isHidden = (self.hiddenTabBatItemButton.isSelected == true ? true : false)
+        self.tabBarSlider02Items[0].isHidden = (self.hiddenTabBatItemButton.isSelected == true ? true : false)
+        
+        self.tabBarSegment01.setTabBarItems(tabBarItemArray: self.segmentTabBarItemArray, isStandAloneView: true)
+        self.tabBarSlider01.setTabBarItems(tabBarItemArray: self.sliderTabBarItemArray, isStandAloneView: true)
+        self.tabBarSlider02.setTabBarItems(tabBarItemArray: self.sliderTabBarItemArray, isStandAloneView: true)
+        self.tabBarChip01.setTabBarItems(tabBarItemArray: self.sliderTabBarItemArray, isStandAloneView: true)
+
+        self.tabBarSegment01ViewController?.hiddenViewPage(tabBarItemArray: self.segmentTabBarItems)
+        self.tabBarSlider01ViewController?.hiddenViewPage(tabBarItemArray: self.tabBarSlider01Items)
+        self.tabBarSlider02ViewController?.hiddenViewPage(tabBarItemArray: self.tabBarSlider02Items)
+    }
+    
+    @objc func changeTitleTabBatItemButtonPressed() {
+        self.changeTitleTabBatItemButton.isSelected.toggle()
+        
+//        self.tabBarView01.changeTabBarButtonTitle(index: 0, title: (self.changeTitleTabBatItemButton.isSelected == true ? "Title 변경" : "원복"))
+//        self.tabBarView02.changeTabBarButtonTitle(index: 3, title: (self.changeTitleTabBatItemButton.isSelected == true ? "Title 변경" : "원복"))
+//        self.tabBarView03.changeTabBarButtonTitle(index: 3, title: (self.changeTitleTabBatItemButton.isSelected == true ? "Title 변경" : "원복"))
+//        self.tabBarView04.changeTabBarButtonTitle(index: 3, title: (self.changeTitleTabBatItemButton.isSelected == true ? "Title 변경" : "원복"))
+//        self.tabBarView05.changeTabBarButtonTitle(index: 3, title: (self.changeTitleTabBatItemButton.isSelected == true ? "Title 변경" : "원복"))
+    }
+}
+
+/*
+ private var tabBarSegment01ViewController: DealiTabBarViewController?
+ private var tabBarSlider01ViewController: DealiTabBarViewController?
+ private var tabBarSlider02ViewController: DealiTabBarViewController?
+ 
+ private let tabBarSegment01 = DealiTabBar_v2.tabBarSegment01()
+ private let tabBarSlider01 = DealiTabBar_v2.tabBarSlider01()
+ private let tabBarSlider02 = DealiTabBar_v2.tabBarSlider02()
+ private let tabBarChip01 = DealiTabBar_v2.tabBarChip01()
+ */
+
+extension TabBarViewController: DealiTabBarViewDelegate {
+    func didSelectTabBar(_ tabbarView: DealiTabBarView, selectedIndex index: Int, showScrollAnimation animation: Bool) {
         print("didSelectTabBar index = \(index)")
     }
 }
