+++ Sources/DealiDesignKit/Components/TabBar/Cell/UIModel/DealiTabBarItemChipStyleCellUIModel.swift
@@ -1,0 +1,28 @@
+//
+//  DealiTabBarItemChipStyleCellUIModel.swift
+//
+//
+//  Created by 이창호 on 11/18/24.
+//
+
+import UIKit
+
+public struct DealiTabBarItemChipStyleCellUIModel {
+    
+    var itemChip: ClickableComponentChip?
+    
+    static func make(chipStyle: DealiTabBarPreset.DealiTabBarSliderChipStyle, tabbarItem: DealiTabBarItem) -> DealiTabBarItemChipStyleCellUIModel {
+        var uiModel = DealiTabBarItemChipStyleCellUIModel()
+        switch chipStyle {
+        case .chipFilledSmall02:
+            uiModel.itemChip = DealiControl.chipFilledSmall02()
+        }
+        
+        uiModel.itemChip?.do {
+            $0.title = tabbarItem.title
+            $0.isUserInteractionEnabled = false
+        }
+        
+        return uiModel
+    }
+}
