+++ Sources/DealiDesignKit/Components/TabBar/DealiTabBarView.swift
@@ -1,8 +1,8 @@
 //
-//  DealiTabBarView.swift
+//  DealiTabBarView_v2.swift
 //  
 //
-//  Created by hoji on 2024/03/26.
+//  Created by 이창호 on 11/4/24.
 //
 
 import UIKit
@@ -26,10 +26,6 @@ public class DealiTabBar {
     public static func tabBarChip01() -> DealiTabBarView {
         return DealiTabBarView(preset: .tabBarChip01)
     }
-    
-    public static func tabBarSlider03() -> DealiTabBarView {
-        return DealiTabBarView(preset: .tabBarSlider03)
-    }
 }
 
 @objc public protocol DealiTabBarViewDelegate {
@@ -37,24 +33,27 @@ public class DealiTabBar {
 }
 
 final public class DealiTabBarView: UIView {
-    
+
     public weak var delegate: DealiTabBarViewDelegate?
 
-    private lazy var contentScrollView = UIScrollView()
-    private let contentStackView = UIStackView()
-    private let selectedUnderLineImageView = UIImageView()
-    private let bottomDividerView = UIView()
+    private let collectionView = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
+    private lazy var selectedLineView = UIView()
+    private lazy var bottomDividerView = UIView()
     
     private var tabBarItemInfoArray: [DealiTabBarItemInfo] = []
     
     /// 해당 TabBarView를 단독으로 사용되면 true / tabBarViewController 와 함께 사용되면 false
     public var isStandAloneView: Bool = false
     
-    private var selectedIndex: Int = 0 {
+    /// TabBar의 구성 및 레이아웃 처리가 정상적으로 완료되었는지에 대한 Bool값(선택된 tab이 center 정렬로 적용하기 위해서는 collectionView width값이 있어야 하는데 then에서 Tabbar item을 구성하게 되면 collectionView width 값이 아직 0.0이라서 레이아웃이 정상적으로 적용되지 않는 이슈로 인해 layoutSubviews 에서 collectionView width값이 세팅되면 그때 다시 레이아웃을 적용하기위해 추가)
+    private var isLayoutInitialized = false
+    
+    private var selectedIndex: Int = -1 {
         didSet {
-            for i in 0..<self.tabBarItemInfoArray.count {
-                self.tabBarItemInfoArray[i].itemSelected = (self.tabBarItemInfoArray[i].itemIndex == self.selectedIndex)
+            for index in 0..<self.tabBarItemInfoArray.count {
+                self.tabBarItemInfoArray[index].itemSelected = (self.tabBarItemInfoArray[index].itemIndex == self.selectedIndex)
             }
+            self.collectionView.reloadData()
         }
     }
     
@@ -73,83 +72,32 @@ final public class DealiTabBarView: UIView {
             $0.height.equalTo(preset.tabBarViewHeight)
         }
         
-        switch preset.style {
-        case .segment:
-            self.addSubview(self.contentStackView)
-            self.contentStackView.then {
-                $0.axis = .horizontal
-                $0.spacing = 0.0
-                $0.distribution = .fillEqually
-                $0.alignment = .center
-            }.snp.makeConstraints {
-                $0.top.bottom.equalToSuperview()
-                $0.left.right.equalToSuperview().inset(preset.tabBarMargin)
+        let layout = UICollectionViewFlowLayout()
+                layout.scrollDirection = .horizontal
+        self.addSubview(self.collectionView)
+        self.collectionView.then {
+            $0.delegate = self
+            $0.dataSource = self
+            if let layout = $0.collectionViewLayout as? UICollectionViewFlowLayout {
+                layout.do {
+                    $0.scrollDirection = .horizontal
+                    $0.minimumLineSpacing = self.preset.itemSpacing
+                    $0.sectionInset = UIEdgeInsets.init(top: 0.0, left: self.preset.tabBarLRMargin, bottom: 0.0, right: self.preset.tabBarLRMargin)
+                }
             }
-            
-            self.addSubview(self.bottomDividerView)
-            self.addSubview(self.selectedUnderLineImageView)
-            
-        default:
-
-            self.addSubview(self.contentScrollView)
-            self.contentScrollView.then {
-                $0.bounces = false
-                $0.clipsToBounds = false
-                $0.showsHorizontalScrollIndicator = false
-                $0.showsVerticalScrollIndicator = false
-            }.snp.makeConstraints {
-                $0.bottom.equalToSuperview()
-                $0.left.right.equalToSuperview().inset(preset.tabBarMargin)
-                $0.height.equalTo(preset.tabBerContentHeight)
-            }
-            
-            let contentView = UIView()
-            self.contentScrollView.addSubview(contentView)
-            contentView.snp.makeConstraints {
-                $0.top.bottom.left.right.equalToSuperview()
-                $0.height.equalToSuperview()
-            }
-            
-            contentView.addSubview(self.contentStackView)
-            self.contentStackView.then {
-                $0.axis = .horizontal
-                $0.spacing = preset.contentSpacing
-                $0.distribution = .equalSpacing
-                $0.alignment = .center
-            }.snp.makeConstraints {
-                $0.top.bottom.left.right.equalToSuperview()
-            }
-            
-            self.addSubview(self.bottomDividerView)
-            self.contentScrollView.addSubview(self.selectedUnderLineImageView)
-            self.bringSubviewToFront(self.contentScrollView)
+            $0.backgroundColor = .white
+            $0.showsVerticalScrollIndicator = false
+            $0.showsHorizontalScrollIndicator = false
+            $0.register(cellClass: DealiTabBarItemTextStyleCell.self)
+            $0.register(cellClass: DealiTabBarItemChipStyleCell.self)
+        }.snp.makeConstraints {
+            $0.left.right.bottom.equalToSuperview()
+            $0.height.equalTo(self.preset.tabBerContentHeight)
+            $0.top.left.bottom.right.equalToSuperview()
         }
         
-        self.bottomDividerView.then {
-            if case .sliderChip(_) = preset.style {
-                $0.isHidden = true
-            } else {
-                $0.isHidden = false
-                $0.backgroundColor = DealiColor.g30
-            }
-        }.snp.makeConstraints {
-            $0.bottom.equalToSuperview()
-            $0.left.right.equalToSuperview().inset(preset.style != .segment ? -preset.tabBarMargin : 0)
-            $0.height.equalTo(1.0)
-        }
-        
-        self.selectedUnderLineImageView.then {
-            if case .sliderChip(_) = preset.style {
-                $0.isHidden = true
-            } else {
-                $0.isHidden = false
-            }
-            $0.backgroundColor = preset.selectedTextColor
-        }.snp.makeConstraints {
-            $0.bottom.equalToSuperview()
-            $0.height.equalTo(2.0)
-            $0.left.equalToSuperview().offset(0.0)
-            $0.width.equalTo(0.0)
+        if preset.style == .segment || preset.style == .slider {
+            self.createSelectedLine()
         }
     }
     
@@ -157,175 +105,150 @@ final public class DealiTabBarView: UIView {
         fatalError("init(coder:) has not been implemented")
     }
     
-    /// TabBar ItemButton click Event
-    @objc func itemPressed(_ sender: UIControl) {
-        self.setSelectedIndex(index: sender.tag, animated: true)
-    }
-    
-    private func setSelectedIndexWithScroll(index: Int) {
-        guard index < self.tabBarItemInfoArray.count else { return }
-        self.selectedIndex = index
-        
-        /// chip을 사용하는 tabBar에서는 따로 underLine 표시되지않기 때문에 chip이 아닌 경우에만 값을 세팅하도록 처리
-        if self.preset.style == .segment || self.preset.style == .sliderButton {
-            self.selectedUnderLineImageView.snp.updateConstraints {
-                $0.left.equalToSuperview().offset(self.tabBarItemInfoArray[self.selectedIndex].contentStartPositionX)
-                $0.width.equalTo(self.tabBarItemInfoArray[self.selectedIndex].contentWidth)
-            }
+    private func createSelectedLine() {
+        self.collectionView.addSubview(self.selectedLineView)
+        self.selectedLineView.then {
+            $0.backgroundColor = preset.selectedTextColor
+        }.snp.makeConstraints {
+            $0.top.equalToSuperview().offset(preset.tabBerContentHeight - 2.0)
+            $0.left.equalToSuperview().offset(0.0)
+            $0.height.equalTo(2.0)
+            $0.width.equalTo(0.0)
         }
         
-        /// TabBar가 단독으로 생성되어 사용되는경우에만 TabBar Button 을 클릭했을 경우 해당 버튼이 화면에 모두 노출되도록 처리
-        if self.preset.style != .segment && self.isStandAloneView {
-            self.moveScrollContentOffset(positionX: self.tabBarItemInfoArray[index].contentStartPositionX, contentWidth: self.tabBarItemInfoArray[index].contentWidth)
+        self.collectionView.backgroundView = UIView()
+        
+        self.collectionView.backgroundView?.addSubview(self.bottomDividerView)
+        self.bottomDividerView.then {
+            $0.backgroundColor = DealiColor.g30
+        }.snp.makeConstraints {
+            $0.bottom.equalToSuperview()
+            $0.left.right.equalToSuperview().inset(-preset.tabBarLRMargin)
+            $0.height.equalTo(1.0)
         }
     }
     
-    public func setSelectedIndex(index: Int, animated: Bool = false, withoutAction: Bool = false) {
+    public override func layoutSubviews() {
+        super.layoutSubviews()
+       
+        if self.collectionView.frame.width > 0.0 && self.collectionView.contentSize.width > 0.0 && self.isLayoutInitialized == false && self.tabBarItemInfoArray.count > 0 {
+
+            self.isLayoutInitialized = true
+            self.setSelectedIndexWithScroll(index: self.selectedIndex, isMoveAnimation: false)
+        }
+    }
+    
+    public func setSelectedIndex(index: Int, animated: Bool = true) {
         self.setSelectedIndexWithScroll(index: index)
-        if !withoutAction {
-            /// tabbar Item button 클릭으로 이벤트 발생시 선택된 Button의 index값을 didSelectTabBarIndex를 통해 전달
-            self.delegate?.didSelectTabBar(self, selectedIndex: self.selectedIndex, showScrollAnimation: animated)
-        }
-    }
-    
-    /// TabBar를 다시 구성하거나할때 기존 TabBar의 정보를 초기화
-    private func clear() {
-        self.tabBarItemInfoArray.forEach { itemInfo in
-            itemInfo.itemButton?.removeFromSuperview()
-            itemInfo.itemChip?.removeFromSuperview()
-        }
-        self.contentStackView.arrangedSubviews.forEach({ $0.removeFromSuperview() })
-        self.tabBarItemInfoArray.removeAll()
+        
+        /// tabbar Item button 클릭으로 이벤트 발생시 선택된 Button의 index값을 didSelectTabBarIndex를 통해 전달
+        self.delegate?.didSelectTabBar(self, selectedIndex: self.selectedIndex, showScrollAnimation: animated)
     }
     
     /// TabBar를 구성할 정보를 받아 TabBar Item Button 생성 및 정보 저장
     public func setTabBarItems(tabBarItemArray: [DealiTabBarItem], maintainContentOffset: Bool = true, startIndex: Int = 0, isStandAloneView: Bool = false) {
-        
         /// 가려지는 tabbar item이 있다면 해당 아이템을 제외하고 TabBarView를 재구성
         let itemArray = tabBarItemArray.filter({ $0.isHidden == false })
         
-        let offset = self.contentScrollView.contentOffset
+        self.tabBarItemInfoArray.removeAll()
         
-        self.clear()
         self.isStandAloneView = isStandAloneView
         for (index, item) in itemArray.enumerated() {
             guard let title = item.title else { continue }
-            var buttonContentWidth = title.size(withAttributes: [.font: self.preset.selectedFont]).width
-            if item.icon != nil, case .sliderButton = self.preset.style  {
-                buttonContentWidth += 16.0
+            var contentWidth = title.size(withAttributes: [.font: self.preset.selectedFont]).width
+            if let _ = item.icon {
+                contentWidth += 16.0 + 2.0
             }
             
             switch self.preset.style {
-            case .sliderChip(let chipStyle):
-                switch chipStyle {
-                case .chipFilledSmall02:
-                    let itemChip = DealiControl.chipFilledSmall02()
-                    self.contentStackView.addArrangedSubview(itemChip)
-                    itemChip.then {
-                        $0.title = title
-                        $0.addTarget(self, action: #selector(itemPressed(_:)), for: .touchUpInside)
-                        $0.tag = index
-                        $0.isSelected = (index == self.selectedIndex)
-                    }.snp.makeConstraints {
-                        $0.centerY.equalToSuperview()
-                    }
-                    
-                    var itemInfo = DealiTabBarItemInfo()
-                    itemInfo.itemIndex = index
-                    itemInfo.itemChip = itemChip
-                    itemInfo.itemSelected = (index == self.selectedIndex)
-                    
-                    self.tabBarItemInfoArray.append(itemInfo)
-                }
-                
-            default:
-                
-                let itemButton = DealiTabBarItemButton()
-                self.contentStackView.addArrangedSubview(itemButton)
-                itemButton.then {
-                    $0.configure(uiModel: DealiTabBarItemButtonUIModel.make(preset: self.preset, tabbarItem: item))
-                    $0.tag = index
-                    $0.isSelected = (index == self.selectedIndex)
-                    $0.addTarget(self, action: #selector(itemPressed(_:)), for: .touchUpInside)
-                }.snp.makeConstraints {
-                    $0.top.bottom.equalToSuperview()
-                    
-                    if case .sliderButton = self.preset.style {
-                        $0.width.equalTo(buttonContentWidth + (self.preset.contentButtonPadding * 2.0))
-                    }
-                }
-            
+            case .segment, .slider:
                 var itemInfo = DealiTabBarItemInfo()
                 itemInfo.itemIndex = index
-                itemInfo.itemButton = itemButton
+                itemInfo.itemTextCellUIModel = DealiTabBarItemTextStyleCellUIModel.make(preset: self.preset, tabbarItem: item)
+                itemInfo.contentWidth = contentWidth
+                
+                if case .segment = self.preset.style {
+                    itemInfo.containerWidth = ((UIScreen.main.bounds.size.width - (self.preset.tabBarLRMargin * 2.0)) / CGFloat(itemArray.count))
+                    itemInfo.contentWidth = itemInfo.containerWidth
+                } else {
+                    itemInfo.containerWidth = contentWidth + (self.preset.itemLRPadding * 2.0)
+                    itemInfo.contentWidth = contentWidth
+                }
+                
+                self.tabBarItemInfoArray.append(itemInfo)
+                
+            case .sliderChip(let chipStyle):
+                var itemInfo = DealiTabBarItemInfo()
+                itemInfo.itemIndex = index
+                itemInfo.itemChipCellUIModel = DealiTabBarItemChipStyleCellUIModel.make(chipStyle: chipStyle, tabbarItem: item)
                 itemInfo.itemSelected = (index == self.selectedIndex)
-                itemInfo.contentWidth = buttonContentWidth
+                if let itemChip = itemInfo.itemChipCellUIModel?.itemChip {
+                    itemInfo.containerWidth = (itemChip.fixedSize.width)
+                    itemInfo.contentWidth = (itemChip.fixedSize.width)
+                }
                 
                 self.tabBarItemInfoArray.append(itemInfo)
             }
         }
         
-        self.updateTabBarItemInfo()
+        var selectedIndex = self.selectedIndex
         
-        self.contentScrollView.setContentOffset((maintainContentOffset ? offset : .zero), animated: false)
-        self.setSelectedIndexWithScroll(index: startIndex)
-    }
-    
-    /// 특정 index에 위치한 tabbaritem의 title 변경 처리
-    public func changeTabBarButtonTitle(index: Int, title: String) {
-        guard index < self.tabBarItemInfoArray.count else { return }
-        
-        if case .sliderChip(_) = self.preset.style {
-            self.tabBarItemInfoArray[index].itemChip?.title = title
+        if selectedIndex == -1 || (startIndex > 0 && startIndex < self.tabBarItemInfoArray.count) {
+            selectedIndex = startIndex
         } else {
-            if var uiModel = self.tabBarItemInfoArray[index].itemButton?.uiModel {
-                let contentWidth = title.size(withAttributes: [.font: self.preset.selectedFont]).width
-                uiModel.title = title
-                self.tabBarItemInfoArray[index].itemButton?.configure(uiModel: uiModel)
-                
-                if case .sliderButton = self.preset.style {
-                    self.tabBarItemInfoArray[index].itemButton?.snp.updateConstraints {
-                        $0.width.equalTo(contentWidth + (self.preset.contentButtonPadding * 2))
-                    }
-                }
+            if selectedIndex >= self.tabBarItemInfoArray.count {
+                selectedIndex = (self.tabBarItemInfoArray.count - 1)
             }
         }
         
-        self.updateTabBarItemInfo()
+        self.selectedIndex = selectedIndex
+        self.updateTabBarItemPositions()
+        
     }
     
-    /// tabbar가 생성되거나 tabbar에 구성된 item의 정보가 변경되었을경우 해당 item 의 최종 width, position X 값을 갱신
-    private func updateTabBarItemInfo() {
+    /// tabbar가 생성되거나 tabbar에 구성된 item의 정보가 변경되었을경우 해당 item 의 position X 값을 갱신 및 세팅
+    private func updateTabBarItemPositions() {
         self.layoutIfNeeded()
         for index in 0..<self.tabBarItemInfoArray.count {
-            if let itemButton = self.tabBarItemInfoArray[index].itemButton {
-                if self.preset.style == .segment {
-                    self.tabBarItemInfoArray[index].contentWidth = ((UIScreen.main.bounds.size.width - (self.preset.tabBarMargin * 2.0)) / CGFloat(self.tabBarItemInfoArray.count))
-                    self.tabBarItemInfoArray[index].contentStartPositionX = self.preset.tabBarMargin + (self.tabBarItemInfoArray[index].contentWidth * CGFloat(index))
-                } else {
-                    self.tabBarItemInfoArray[index].contentStartPositionX = itemButton.frame.origin.x + self.preset.contentButtonPadding
-                    self.tabBarItemInfoArray[index].contentWidth = itemButton.bounds.size.width - (self.preset.contentButtonPadding * 2) + (itemButton.uiModel?.iconSize != nil ? 2.0 : 0.0) // 아이콘이 있는 경우 2px 추가 (디자인팀 요청)
+            
+            let indexPath = IndexPath(item: index, section: 0)
+            
+            if let attributes = collectionView.layoutAttributesForItem(at: indexPath) {
+                let cellXPosition = attributes.frame.origin.x
+                if case .segment = self.preset.style {
+                    self.tabBarItemInfoArray[index].contentPositionX = cellXPosition
+                } else if case .slider = self.preset.style {
+                    self.tabBarItemInfoArray[index].contentPositionX = cellXPosition + self.preset.itemLRPadding
+                } else if case .sliderChip(_) = self.preset.style {
+                    self.tabBarItemInfoArray[index].contentPositionX = cellXPosition
                 }
-            } else if let itemChip = self.tabBarItemInfoArray[index].itemChip {
-                self.tabBarItemInfoArray[index].contentStartPositionX = (itemChip.frame.origin.x)
-                self.tabBarItemInfoArray[index].contentWidth = itemChip.frame.size.width
             }
         }
         
-        if self.preset.style == .segment || self.preset.style == .sliderButton {
-            self.selectedUnderLineImageView.snp.updateConstraints {
-                $0.left.equalToSuperview().offset(self.tabBarItemInfoArray[self.selectedIndex].contentStartPositionX)
-                $0.width.equalTo(self.tabBarItemInfoArray[self.selectedIndex].contentWidth)
-            }
+        if self.isLayoutInitialized == true {
+            self.setSelectedIndexWithScroll(index: self.selectedIndex)
         }
     }
     
-    public func showTabBarItemBadge(index: Int, shouldShowBadge: Bool) {
-        guard index < self.tabBarItemInfoArray.count, let itemButton =  self.tabBarItemInfoArray[index].itemButton else { return }
-        if self.preset.style == .segment || self.preset.style == .sliderButton {
-            if var tabBarItemButtonUIModel = itemButton.uiModel {
-                tabBarItemButtonUIModel.shouldExposeNewBadge = shouldShowBadge
-                self.tabBarItemInfoArray[index].itemButton?.configure(uiModel: tabBarItemButtonUIModel)
+    private func setSelectedIndexWithScroll(index: Int, isMoveAnimation: Bool = true) {
+        guard index < self.tabBarItemInfoArray.count else { return }
+        self.selectedIndex = index
+
+        
+        /// TabBar가 단독으로 생성되어 사용되는경우에만 TabBar Button 을 클릭했을 경우 해당 버튼이 화면에 모두 노출되도록 처리
+        if self.preset.style != .segment && self.isStandAloneView {
+            self.moveScrollContentOffset(positionX: self.tabBarItemInfoArray[index].contentPositionX, contentWidth: self.tabBarItemInfoArray[index].contentWidth, isMoveAnimation: isMoveAnimation)
+        }
+        
+        self.updateSelectedLine(width: self.tabBarItemInfoArray[self.selectedIndex].contentWidth, positionX: self.tabBarItemInfoArray[self.selectedIndex].contentPositionX)
+    }
+    
+    /// chip을 사용하는 tabBar에서는 따로 underLine 표시되지않기 때문에 chip이 아닌 경우에만 값을 세팅하도록 처리
+    private func updateSelectedLine(width: CGFloat, positionX: CGFloat) {
+        if self.preset.style == .segment || self.preset.style == .slider {
+            self.selectedLineView.snp.updateConstraints {
+                $0.left.equalToSuperview().offset(positionX)
+                $0.width.equalTo(width)
             }
         }
     }
@@ -346,58 +269,112 @@ final public class DealiTabBarView: UIView {
 
         if preIdx < 0 {
             if let item = self.tabBarItemInfoArray.first {
-                positionX = item.contentStartPositionX
+                positionX = item.contentPositionX
                 contentWidth = item.contentWidth
             }
         } else if nexIdx >= self.tabBarItemInfoArray.count {
             if let item = self.tabBarItemInfoArray.last {
-                positionX = item.contentStartPositionX
+                positionX = item.contentPositionX
                 contentWidth = item.contentWidth
             }
         } else {
             let preItem = self.tabBarItemInfoArray[preIdx]
             let nexItem = self.tabBarItemInfoArray[nexIdx]
-            positionX = preItem.contentStartPositionX + (nexItem.contentStartPositionX - preItem.contentStartPositionX) * calc
+            positionX = preItem.contentPositionX + (nexItem.contentPositionX - preItem.contentPositionX) * calc
             contentWidth = (preItem.contentWidth) + ((nexItem.contentWidth) - (preItem.contentWidth)) * calc
         }
         
         self.selectedIndex = page
         
-        if self.preset.style == .segment || self.preset.style == .sliderButton {
-            self.selectedUnderLineImageView.snp.updateConstraints {
-                $0.width.equalTo(contentWidth)
-                $0.left.equalToSuperview().offset(positionX)
-            }
-        }
-
-        if self.preset.style == .segment {
-            return
+        self.updateSelectedLine(width: contentWidth, positionX: positionX)
+        
+        if self.preset.style != .segment {
+            self.moveScrollContentOffset(positionX: positionX, contentWidth: contentWidth)
         }
         
-        self.moveScrollContentOffset(positionX: positionX, contentWidth: contentWidth)
     }
     
     /// tabbar Item button을 클릭하거나 ViewController에서 스크롤이 발생했을경우 해당 선택된 tabbar Item Button이 화면에 노출되도록 offset 변경
-    private func moveScrollContentOffset(positionX: CGFloat, contentWidth: CGFloat) {
+    private func moveScrollContentOffset(positionX: CGFloat, contentWidth: CGFloat, isMoveAnimation: Bool = false) {
         var offset: CGFloat = -1
-
+        let centerOffsetX = (self.collectionView.frame.width / 2)
+        offset = positionX - centerOffsetX + (contentWidth / 2)
+        offset = max(offset, 0)
+        let maxOffsetX = self.collectionView.contentSize.width - self.collectionView.frame.width
+        offset = min(offset, maxOffsetX)
+        
+        if offset >= 0 {
+            self.collectionView.setContentOffset(CGPoint(x: offset, y: self.collectionView.contentOffset.y), animated: isMoveAnimation)
+        }
+    }
+    
+    
+    /// 특정 index에 위치한 tabbaritem의 title 변경 처리
+    public func changeTabBarButtonTitle(index: Int, title: String) {
+        guard index < self.tabBarItemInfoArray.count else { return }
+        
         if case .sliderChip(_) = self.preset.style {
-            if positionX < self.contentScrollView.contentOffset.x || self.contentScrollView.frame.width <= 0 {
-                offset = positionX
-            } else if (positionX + contentWidth) > self.contentScrollView.contentOffset.x + self.contentScrollView.frame.width {
-                offset = (positionX + contentWidth) - self.contentScrollView.frame.width
+            self.tabBarItemInfoArray[index].itemChipCellUIModel?.itemChip?.title = title
+            if let itemChip = self.tabBarItemInfoArray[index].itemChipCellUIModel?.itemChip {
+                self.tabBarItemInfoArray[index].contentWidth = (itemChip.fixedSize.width)
+                self.tabBarItemInfoArray[index].containerWidth = (itemChip.fixedSize.width)
             }
         } else {
-            if (positionX - self.preset.contentButtonPadding) < self.contentScrollView.contentOffset.x || self.contentScrollView.frame.width <= 0 {
-                offset = (positionX - self.preset.contentButtonPadding)
-            } else if (positionX + contentWidth + self.preset.contentButtonPadding) > self.contentScrollView.contentOffset.x + self.contentScrollView.frame.width {
-                offset = (positionX + contentWidth + self.preset.contentButtonPadding) - self.contentScrollView.frame.width
+            if var uiModel = self.tabBarItemInfoArray[index].itemTextCellUIModel {
+                var contentWidth = title.size(withAttributes: [.font: self.preset.selectedFont]).width
+                if let _ = uiModel.iconURL {
+                    contentWidth += 16.0 + 2.0
+                }
+                self.tabBarItemInfoArray[index].contentWidth = contentWidth
+                self.tabBarItemInfoArray[index].containerWidth = contentWidth + (self.preset.itemLRPadding * 2.0)
+                
+                uiModel.title = title
             }
         }
+        self.collectionView.reloadData()
+        self.updateTabBarItemPositions()
+    }
 
-        if offset >= 0 {
-            self.contentScrollView.setContentOffset(CGPoint(x: offset, y: self.contentScrollView.contentOffset.y), animated: true)
+    public func showTabBarItemBadge(index: Int, shouldShowBadge: Bool) {
+        guard index < self.tabBarItemInfoArray.count else { return }
+        if self.preset.style == .segment || self.preset.style == .slider {
+            print("showTabBarItemBadge segment / slider")
+            self.tabBarItemInfoArray[index].itemTextCellUIModel?.shouldExposeNewBadge = shouldShowBadge
+            self.collectionView.reloadData()
         }
+    }
+}
+
+extension DealiTabBarView: UICollectionViewDataSource, UICollectionViewDelegate {
+    public func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return self.tabBarItemInfoArray.count
+    }
+    
+    public func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+             
+        if case .sliderChip(_) = self.preset.style {
+            let cell = collectionView.dequeueReusableCell(cellClass: DealiTabBarItemChipStyleCell.self, indexPath: indexPath)
+            cell.configure(uiModel: self.tabBarItemInfoArray[indexPath.item].itemChipCellUIModel)
+            return cell
+        } else {
+            let cell = collectionView.dequeueReusableCell(cellClass: DealiTabBarItemTextStyleCell.self, indexPath: indexPath)
+            cell.configure(uiModel: self.tabBarItemInfoArray[indexPath.item].itemTextCellUIModel)
+            return cell
+        }
+        
+    }
+    
+    public func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        if indexPath.item == self.selectedIndex {
+            return
+        }
+        self.setSelectedIndex(index: indexPath.item)
+    }
+}
+
+extension DealiTabBarView: UICollectionViewDelegateFlowLayout {
+    public func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: self.tabBarItemInfoArray[indexPath.item].containerWidth, height: self.preset.tabBarViewHeight)
     }
 }
 
@@ -433,19 +410,22 @@ public struct DealiTabBarIcon {
 }
 
 struct DealiTabBarItemInfo {
+    /// View width 정보
+    var containerWidth: CGFloat = 0.0
+    /// View 안에 들어가는 content width 정보
     var contentWidth: CGFloat = 0.0
-    var contentStartPositionX: CGFloat = 0.0
+    var contentPositionX: CGFloat = 0.0
     var itemIndex: Int = 0
-    var itemButton: DealiTabBarItemButton?
-    var itemChip: ClickableComponentChip?
+    var itemTextCellUIModel: DealiTabBarItemTextStyleCellUIModel?
+    var itemChipCellUIModel: DealiTabBarItemChipStyleCellUIModel?
     var itemSelected: Bool = false {
         didSet {
-            if let itemChip = self.itemChip {
-                itemChip.isSelected = self.itemSelected
+            if let _ = self.itemTextCellUIModel {
+                self.itemTextCellUIModel?.isSelected = self.itemSelected
             }
             
-            if let itemButton = self.itemButton {
-                itemButton.isSelected = self.itemSelected
+            if let itemCellUIModel = self.itemChipCellUIModel, let itemChip = itemCellUIModel.itemChip {
+                itemChip.isSelected = self.itemSelected
             }
         }
     }
